Shrink RunManager admission critical section
